Split memory tests to separate test class
Allows integration tests for features to be run separately.
