+++ tests/IntegrationTest.php
@@ -13,13 +13,11 @@ use Phlib\Flysystem\Pdo\PdoAdapter;
  */
 class IntegrationTest extends IntegrationTestCase
 {
-    use MemoryTestTrait;
-
     protected static array $tempFileSize = [
         '00B' => 0,
         '10B' => 10,
         '10K' => 10 * 1024,
-        'xl' => 10 * 1024 * 1024,
+        'l' => 2 * 1024 * 1024,
     ];
 
     private PdoAdapter $adapter;
@@ -221,78 +219,6 @@ class IntegrationTest extends IntegrationTestCase
         static::assertSame($origDataSet, $copyDataSet);
     }
 
-    public function testMemoryUsageOnWritingStream(): void
-    {
-        $filename = static::$tempFiles['xl'];
-        $file = fopen($filename, 'r');
-        $path = '/path/to/file.txt';
-
-        $variation = 1048576; // 1MiB
-        $this->memoryTest(function () use ($path, $file): void {
-            $this->adapter->writeStream($path, $file, $this->emptyConfig);
-        }, $variation);
-    }
-
-    public function testMemoryUsageOnReadingStreamWithBuffering(): void
-    {
-        $config = $this->emptyConfig;
-        if (static::getDbDriverName() === 'mysql') {
-            $config = new Config([
-                'enable_mysql_buffering' => true,
-            ]);
-        }
-        $adapter = new PdoAdapter(static::getTestDbAdapter(), $config);
-
-        $filename = static::$tempFiles['xl'];
-        $file = fopen($filename, 'r');
-        $path = '/path/to/file.txt';
-
-        $adapter->writeStream($path, $file, $this->emptyConfig);
-
-        $variation = 1048576; // 1MiB
-        $this->memoryTest(function () use ($adapter, $path): void {
-            $adapter->readStream($path);
-        }, $variation);
-    }
-
-    public function testMemoryUsageOnReadingStreamWithoutBuffering(): void
-    {
-        if (static::getDbDriverName() !== 'mysql') {
-            static::markTestSkipped('Cannot test buffering on non mysql driver.');
-        }
-
-        $config = new Config([
-            'enable_mysql_buffering' => false,
-        ]);
-        $adapter = new PdoAdapter(static::getTestDbAdapter(), $config);
-
-        $filename = static::$tempFiles['xl'];
-        $file = fopen($filename, 'r');
-        $path = '/path/to/file.txt';
-
-        $adapter->writeStream($path, $file, $this->emptyConfig);
-
-        $variation = 1048576; // 1MiB
-        $this->memoryTest(function () use ($adapter, $path): void {
-            $adapter->readStream($path);
-        }, $variation);
-    }
-
-    public function testMemoryUsageOnUpdateStream(): void
-    {
-        $path = '/path/to/file.txt';
-        $file = fopen(static::$tempFiles['10K'], 'r');
-        $this->adapter->writeStream($path, $file, $this->emptyConfig);
-        fclose($file);
-
-        $file = fopen(static::$tempFiles['xl'], 'r');
-
-        $variation = 1048576; // 1MiB
-        $this->memoryTest(function () use ($path, $file): void {
-            $this->adapter->updateStream($path, $file, $this->emptyConfig);
-        }, $variation);
-    }
-
     /**
      * @dataProvider pathsDataProvider
      */
@@ -382,10 +308,10 @@ class IntegrationTest extends IntegrationTestCase
 
     public function testDeletingFileClearsAllChunks(): void
     {
-        $file = file_get_contents(static::$tempFiles['xl']);
+        $file = file_get_contents(static::$tempFiles['l']);
         $this->adapter->write('/test.txt', $file, $this->emptyConfig);
 
-        static::assertRowCount(8, 'flysystem_chunk');
+        static::assertRowCount(2, 'flysystem_chunk');
         $this->adapter->delete('/test.txt');
         static::assertRowCount(0, 'flysystem_chunk');
     }
@@ -436,7 +362,7 @@ class IntegrationTest extends IntegrationTestCase
     /**
      * @return resource
      */
-    protected function createResource(string $filename)
+    private function createResource(string $filename)
     {
         $handle = fopen($filename, 'r');
         $this->tempHandles[] = $handle;
